% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addColorBar.R
\name{addColorBar}
\alias{addColorBar}
\title{Add a color bar to a plot}
\usage{
addColorBar(
  zlim,
  cols,
  bar_width = 0.035,
  bar_height = 0.3,
  x_pad = 0.04,
  y_pad = 0.05,
  lab_name = NULL,
  lab_cex = 0.9,
  tick_cex = 0.9,
  tick_num = 5,
  tick_digits = 1
)
}
\arguments{
\item{zlim}{\code{Numeric} values giving the min and max values of the color bar.}

\item{cols}{A \code{character} vector of colors to display on the bar.}

\item{bar_width, }{bar_height \code{Numeric} values between 0 and 1, denoting the size of the bar relative to the size of the plot.}

\item{x_pad, }{y_pad \code{Numeric} values between 0 and 1, denoting the position of the bar relative to the size of the plot.}

\item{lab_name}{(optional) A \code{character} string to display on the side of the bar.}

\item{lab_cex}{(optional) A \code{numeric} value denoting the relative size of the label text.}

\item{tick_cex, }{tick_num, tick digits (optional) \code{numeric} values denoting the relative size, number and digits to display for the values alongside the bar.}

\item{lab_namee}{(optional) A \code{character} string to display on the side of the bar.}
}
\value{
This function adds a color bar to an existing plot.
}
\description{
The \code{addColorBar} function adds a color bar to an already existing plot.
It is used to e.g. plot a continuous legend on a map.
The position and size of the bar are given by its position relative to the size of the plot (starting in the bottom left corner).
}
\examples{
graphics::par(mar = c(0,0,0,0))
plot(sf::st_geometry(COUNTRIES), border = NA, col = "gray80")
zlim <-  c(0,1)
cols <- hcl.colors(100, "Viridis")
addColorBar( zlim = zlim, cols = cols, y_pad = 0.5, x_pad = 0.15, lab_name = "something", lab_cex = 1.5)

}
\author{
Pierre Dupont
}
