% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addPNG.R
\name{addPNG}
\alias{addPNG}
\title{Add a .png image to a plot}
\usage{
addPNG(x, y, name, size, path = NULL)
}
\arguments{
\item{x, y}{\code{Numeric} vectors of values between 0 and 1, one value for each item in the \code{name} list, denoting the relative position(s) along the x- and y-axes.}

\item{name}{A \code{character} vector to be matched with the name of the image.}

\item{size}{A \code{numeric} of values between 0 and 1, denoting the size of the image relative to the width of the plot.}

\item{path}{(optional) A \code{path} to the directory containing the image file(s).}
}
\value{
This function adds .png images to an existing plot.
}
\description{
The \code{addPNG} function adds one or several .png images to an already existing plot.
The position and size of the image is given by its coordinates relative to the size of the plot.
}
\examples{
graphics::par(mar = c(0,0,0,0))
plot(sf::st_geometry(COUNTRIES), border = NA, col = "gray80")
addPNG(x = c(0.5,0.8), y = c(0.5,0.15), name = c("prrt","bear"), size = c(0.8,0.2))

}
\author{
Pierre Dupont
}
