% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addPopSize.R
\name{addPopSize}
\alias{addPopSize}
\title{Add population size estimates to plot}
\usage{
addPopSize(x, y, labels)
}
\arguments{
\item{x, y}{\code{Numeric} vectors of values between 0 and 1, one value for each item in the \code{text} list, denoting the relative position(s) along the x- and y-axes.}

\item{text}{A named \code{list} containing the population size estimates. 
If names are provided, they will be used to match country flags.}
}
\value{
This function adds population size estimates as text and country flags.
}
\description{
Function to add population size estimates and country flags to an existing plot.
}
\details{
The \code{addPopSize} function uses the outputs from the \code{getDensity} function as input.
}
\examples{
plot(sf::st_geometry(COUNTRIES), border = NA, col = "gray80")

labels <- list(NOR = c(90,123), SWEden = c(20,40))
x <- c(0.1,0.6)
y <- c(0.7,0.3)
addPopSize(x,y,labels)

}
\author{
Pierre Dupont
}
