% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addScale.R
\name{addScale}
\alias{addScale}
\title{Add a scale to a map plot}
\usage{
addScale(x, y, size, dir = "v")
}
\arguments{
\item{x, y}{\code{Numeric} vectors of values between 0 and 1, one value for each item in the \code{name} list, denoting the relative position of the center of the scale along the x- and y-axes.}

\item{size}{A \code{numeric} of values between 0 and 1, denoting the size of the image relative to the width of the plot.}

\item{dir}{A \code{character} denoting the orientation of the scale; \code{"v"} (the default value).}
}
\value{
This function adds a scale (in km) to an existing map.
}
\description{
The \code{addScale} function adds a scale with text to a map.
The position, size and orientation of the scale are given.
}
\examples{
graphics::par(mar = c(0,0,0,0))
plot(sf::st_geometry(COUNTRIES), border = NA, col = "gray80")
addScale(x = 0.8, y = 0.2, size = 350000, dir = "h")

}
\author{
Pierre Dupont
}
