% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addSnippets.R
\name{addSnippets}
\alias{addSnippets}
\title{Add R snippets}
\usage{
addSnippets(path = getwd(), snippetsFile = "snippets.R")
}
\arguments{
\item{path}{A path indicating where to search for the snippet file.}

\item{snippetsFile}{A \code{character} object with the name of thefile containing the snippets definitions.}
}
\value{
boolean invisible(FALSE) if nothing was added, invisible(TRUE) if snippet definitions were added
}
\description{
\code{addSnippets} copies all (missing) snippet definitions in 
\code{snippetsFile} to the RStudio's user snippet location.
(adapted for RovQuant from :
https://stackoverflow.com/questions/62220596/can-r-packages-add-code-snippets-to-users-snippet-files)
}
\examples{
\dontrun{addSnippets()}

}
