% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AssignDetectors.R
\name{AssignDetectors}
\alias{AssignDetectors}
\title{Detectors Assignment Function}
\usage{
AssignDetectors(myData, myDetectors, mysubDetectors = NULL, radius = 20000)
}
\arguments{
\item{myData}{A \code{sf} object containing the detection data}

\item{mysubDetectors}{Optional; a \code{list} of sf objects containing the subdetectors locations (for PAB models only)}

\item{radius}{a \code{numeric} value denoting the maximum radius to consider for detector assignment. 
Any detection further than \code{radius}m from any detector will not be assigned to any detector. Instead it will get a 'NA'.}

\item{myDetector}{A \code{list} of sf objects containing the detector locations}
}
\value{
A \code{sf} object with the x and y locations of the
 different detections with associated detector's Id
}
\description{
\code{AssignDetectors} identifies and assigns the closest detector to each detection in \code{myData}.
}
