% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignSearchTracks.R
\name{assignSearchtTracks}
\alias{assignSearchtTracks}
\title{Assign DNA samples to search tracks bite-size NIMBLE MCMC algorithms}
\usage{
assignSearchtTracks(data, tracks, dist = 500, progress.bar = T)
}
\arguments{
\item{data}{A \code{sf} object containing the detection data}

\item{dist}{a \code{numeric} value denoting the maximum distance to consider for tracks assignment. 
Any detection further than \code{dist} m from any track will not be assigned to any track. Instead it will get a 'NA'.}

\item{progress.bar}{a \code{logical}, whether to display a progress bar or not.}

\item{detectors}{A \code{tracks} of sf objects containing the search tracks}
}
\value{
A \code{sf} object with the x and y locations of the different detections with associated tracks Id
}
\description{
the \code{assignSearchtTracks} function is a R function used to match RovBase
DNA samples to the corresponding search track. Matching is based on date 
(exact match) and distance (within a given threshold).
samples on the fly.
}
\author{
Pierre Dupont
}
