% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkInputValidity.R
\name{checkInputValidity}
\alias{checkInputValidity}
\alias{checkInputValidity_lite}
\title{Input checking function}
\usage{
checkInputValidity(
  id = NULL,
  s,
  y,
  z,
  lowerHabCoords,
  upperHabCoords,
  localHabWindowIndices = NULL,
  numLocalHabWindows = NULL,
  trapCoords,
  localTrapsIndices = NULL,
  localTrapsNum = NULL,
  resizeFactor = 1,
  habitatGrid,
  habitatGridLocal = NULL,
  silent = FALSE,
  printReport = FALSE,
  pathReport = NULL
)

checkInputValidity_lite(
  id = NULL,
  s,
  y,
  z,
  lowerHabCoords,
  upperHabCoords,
  localHabWindowIndices = NULL,
  numLocalHabWindows = NULL,
  trapCoords,
  localTrapsIndices = NULL,
  localTrapsNum = NULL,
  resizeFactor = 1,
  habitatGrid,
  habitatGridLocal = NULL
)
}
\arguments{
\item{s}{array of activity center coordinates.}

\item{y}{detection histories.}

\item{z}{matrix of individual states.}

\item{lowerHabCoords}{lower coordinates of habitat grid cells.}

\item{upperHabCoords}{upper coordinates of habitat grid cells.}

\item{localHabWindowIndices}{matrix of IDs of the local habitat grid cells.}

\item{numLocalHabWindows}{vector of numbers of local habitat grid cells.}

\item{trapCoords}{matrix of detector coordinates}

\item{localTrapsIndices}{matrix of IDs of the local detectors.}

\item{localTrapsNum}{vector of numbers of local detectors.}

\item{resizeFactor}{default value= 1.}

\item{habitatGrid}{matrix of habitat grid cell IDs.}

\item{habitatGridLocal}{matrix of resized habitat grid cell IDs.}

\item{silent}{Logical; whether to print out outputs or not}

\item{printReport}{Logical; whether to print out a .pdf report or not}

\item{pathReport}{(optional) path to print the report}

\item{data}{A \code{sf} object containing the detection data}

\item{detectors}{A \code{list} of sf objects containing the detector locations}

\item{subDetectors}{Optional; a \code{list} of sf objects containing the subdetectors locations (for PAB models only)}

\item{radius}{a \code{numeric} value denoting the maximum radius to consider for detector assignment. 
Any detection further than \code{radius}m from any detector will not be assigned to any detector. Instead it will get a 'NA'.
   id = NULL,}
}
\value{
A \code{.pdf} report with potentially problematic individual detections and/or movements.
}
\description{
\code{checkInputValidity} checks the validity of the nimbleSCR input for OPSCR models.
}
