% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rovquantR_data.R
\docType{data}
\name{COUNTIES}
\alias{COUNTIES}
\title{COUNTIES}
\format{
## `COUNTIES`
A simple feature collection with 40 features and 1 field
}
\source{
<https://www.nmbu.no/en/research/projects/rovquant>
}
\usage{
COUNTIES
}
\description{
These data sets describe Scandinavian administrative borders.

`COUNTIES`, `REGIONS`, `COUNTRIES`, `COUNTRIESWaterHumans` and `GLOBALMAP` are simple feature (sf) objects.
`habitatRasterResolution` and `habitatRasters` are raster objects.
}
\examples{
data(COUNTIES)
}
\keyword{datasets}
