% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rovquantR_data.R
\docType{data}
\name{COUNTRIES}
\alias{COUNTRIES}
\title{COUNTRIES}
\format{
## `COUNTRIES`
A simple feature collection with 2 features and 1 field
}
\source{
<https://www.nmbu.no/en/research/projects/rovquant>
}
\usage{
COUNTRIES
}
\description{
These data sets describe Scandinavian countries.

`COUNTIES`, `REGIONS`, `COUNTRIES`, `COUNTRIESWaterHumans` and `GLOBALMAP` are simple feature (sf) objects.
`habitatRasterResolution` and `habitatRasters` are raster objects.
}
\examples{
data(COUNTRIES)
}
\keyword{datasets}
