% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rovquantR_data.R
\docType{data}
\name{COUNTRIESWaterHumans}
\alias{COUNTRIESWaterHumans}
\title{COUNTRIESWaterHumans}
\format{
## `COUNTRIESWaterHumans`
A Simple feature collection with 56 features and 10 fields
}
\source{
<https://www.nmbu.no/en/research/projects/rovquant>
}
\usage{
COUNTRIESWaterHumans
}
\description{
These data sets describe Scandinavian borders.

`COUNTIES`, `REGIONS`, `COUNTRIES`, `COUNTRIESWaterHumans` and `GLOBALMAP` are simple feature (sf) objects.
`habitatRasterResolution` and `habitatRasters` are raster objects.
}
\examples{
data(COUNTRIESWaterHumans)
}
\keyword{datasets}
