% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbin_LESS_Cached_MultipleCovResponse.R
\name{dbin_LESS_Cached_MultipleCovResponse}
\alias{dbin_LESS_Cached_MultipleCovResponse}
\title{NIMBLE custom distribution for faster SCR model runs.}
\usage{
dbin_LESS_Cached_MultipleCovResponse(
  x,
  sxy,
  sigma,
  nbDetections,
  yDets,
  detector.xy,
  trials,
  detectorIndex,
  nDetectorsLESS,
  ResizeFactor = 1,
  maxNBDets,
  habitatID,
  indicator = 1,
  p0State,
  detCountries,
  detCov,
  betaCov,
  BetaResponse,
  detResponse,
  log = 0
)
}
\arguments{
\item{x}{\code{Vector} of length n.detectors containing detection/non detections}
}
\description{
\code{dbin_LESS_Cached_OneCov} returns the likelihood of a given individual & spatial binary detection history y[i,1:n.detectors]
}
\examples{
y[i,1:n.detectors,t] ~ dbern_LESS(p0, sigma, d2[i,1:n.detectors,t], maxDist,  z[i,t]==2)
}
