% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{extractUniquePositiveValues}
\alias{extractUniquePositiveValues}
\title{Mode  calculation}
\usage{
extractUniquePositiveValues(matrix)
}
\arguments{
\item{x}{A numeric vector whose sample quantiles are wanted.}

\item{narm}{A logical, whether to remove NA values before calculation or not.}
}
\description{
This function returns the sample mode.
}
