% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{GetDensity}
\alias{GetDensity}
\alias{quantileCpp}
\alias{fastIntMode}
\alias{extractUniquePositiveValues}
\alias{createTransitionMatrix}
\alias{GetSpaceUse}
\alias{GetDetectability_normal}
\alias{GetDetectability_mean}
\alias{GetTransitions}
\title{MCMC output post-processing}
\usage{
quantileCpp(x, q)

fastIntMode(x, narm = FALSE)

extractUniquePositiveValues(matrix)

createTransitionMatrix(values)

GetDensity(
  sx,
  sy,
  z,
  IDmx,
  aliveStates,
  regionID,
  probs = as.numeric(c(0.025, 0.975)),
  display_progress = TRUE,
  returnPosteriorCells = TRUE
)

GetSpaceUse(
  sx,
  sy,
  z,
  sigma,
  habitatxy,
  aliveStates,
  regionID,
  probs = as.numeric(c(0.025, 0.975)),
  display_progress = TRUE,
  returnPosteriorCells = TRUE
)

GetDetectability_normal(
  p0,
  sigma,
  habitatxy,
  detectorxy,
  regionID,
  probs = as.numeric(c(0.025, 0.975)),
  localDist = 100,
  display_progress = TRUE,
  returnPosteriorCells = FALSE
)

GetDetectability_mean(
  p0,
  sigma,
  habitatxy,
  detectorxy,
  display_progress = TRUE
)

GetTransitions(
  sx1,
  sy1,
  z1,
  sx2,
  sy2,
  z2,
  stateFrom,
  stateTo,
  rgmx,
  probs = as.numeric(c(0.025, 0.975)),
  display_progress = TRUE
)
}
\arguments{
\item{x}{A numeric vector whose sample quantiles (\code{quantileCpp}) or mode (\code{fastIntMode}) are wanted.}

\item{q}{A numeric vector with the values of the quantiles to be returned.}

\item{narm}{A logical, whether to remove NA values before calculation or not.}

\item{matrix}{A numeric matrix}

\item{values}{A numeric vector}

\item{sx}{A numeric matrix containing posterior samples of individual activity center's x-coordinates.}

\item{sy}{A numeric matrix containing posterior samples of individual activity center's x-coordinates.}

\item{z}{A numeric matrix containing posterior samples of individual states.}

\item{IDmx}{A numeric matrix containing the IDs of the habitat cells.}

\item{aliveStates}{A numeric vector of the alive states.}

\item{regionID}{A binary matrix with dimensions (number of regions)*(number of habitat cells), denoting whether a given cell belongs the region (1) or not (0).}

\item{probs}{numeric vector with the values of the quantiles to be returned for the confidence interval.}

\item{display_progress}{A logical, whether to display progress bar or not.}

\item{returnPosteriorCells}{A logical, whether to return posterior values for each cell or not.}

\item{sigma}{A numeric matrix containing posterior samples of individual sigma values.}

\item{habitatxy}{A numeric matrix containing the coordinates of all habitat cells.}

\item{p0}{A numeric matrix containing posterior samples of individual baseline detection probabilities.}

\item{detectorxy}{A numeric matrix containing the coordinates of all detectors.}

\item{localDist}{A numeric value}

\item{sx1, sx2, sy1, sy2}{Numeric matrices of posterior x- and y-coordinates.}

\item{z1, z2}{Numeric matrices of posterior x- and y-coordinates.}

\item{stateFrom, stateTo}{departure and arrival states.}

\item{rgmx}{A numeric matrix}
}
\value{
Return_Description
}
\description{
This set of \code{C++} functions are meant to facilitate the post-processing of SCR and OPSCR posterior samples in order to provide relevant information and derived quantities.
}
\details{
The \code{GetDensity} function extracts density estimates for a set of grid cells based on individual posterior activity center locations.
The \code{GetSpaceUse} function extracts density estimates for a set of grid cells based on individual posterior space use distribution.
The \code{GetDetectability_normal} function calculates detectability, i.e. the probability for an individual to be detected at least once for a set of grid cells based on posterior baseline detection probabilities and scale parameters. This function assumes a half-normal detection function.
The \code{GetDetectability_mean} function is a fast alternative of the \code{GetDetectability_normal} based on the posterior means of the parameters instead of the full posterior MCMC.
}
