% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDensityInput.R
\name{getDensityInput}
\alias{getDensityInput}
\title{Density objects preparation}
\usage{
getDensityInput(regions, habitat, s = NULL, plot.check = TRUE)
}
\arguments{
\item{regions}{Raster of the region(s) for which density will be calculated, 
at the desired resolution.}

\item{habitat}{Raster of the habitat (as used in the SCR model).}

\item{s}{Array of MCMC samples for individual activity center x- and y-coordinates 
(NB: make sure these are in the same projection than the \code{habitat} and \code{regions}
 objects).}

\item{plot.check}{A visualization option (if TRUE); 
displays the raster of the different regions that will be used to extract density,
as well as potential mismatches between the \code{habitat} and \code{regions} rasters.}
}
\value{
This function returns a list of objects:
\itemize{
\item regions.r : raster of the regions for which density will be calculated, with the desired resolution.
\item habitat.id: matrix of habitat grid cell IDs (NA for non-habitat; 0 for habitat not in the \code{regions} considered). 
\item regions.rgmx: matrix of dimensions number of regions * number of habitat cells containing
 binary indicators denoting to which region each habitat cell belongs.
\item sx: matrix of rescaled MCMC samples of x-coordinates of individual activity centers for use in the Rcpp \code{getDensity} function.  
\item sy: matrix of rescaled MCMC samples of y-coordinates of individual activity centers for use in the Rcpp \code{getDensity} function.
}
}
\description{
R utility function to prepare input objects for the \code{getDensity} function.
}
\details{
The \code{getDensityInput} function is used in advance of density extraction with function \code{getDensity}.
}
\author{
Pierre Dupont
}
