% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSInits.R
\name{getSInits}
\alias{getSInits}
\title{Generation of activity centers for SCR and OPSCR initial values.}
\usage{
getSInits(
  AllDetections = AllDetections,
  Id.vector = Id.vector,
  idAugmented = idAugmented,
  lowerCoords = lowerCoords,
  upperCoords = upperCoords,
  habitatGrid = habitatGrid,
  intensity = NULL,
  sd = 4,
  movementMethod = "dbernppACmovement_normal"
)
}
\arguments{
\item{AllDetections}{a 2- or 3-dimensional array of individual detections. 
This argument can be provided in two formats: 
(i) with the \emph{y} object as returned by \code{\link{getSparseY}} 
(ii) with the \emph{yCombined} object as returned by \code{\link{getSparseY}}.
The \emph{yCombined} object combines \emph{detNums}, \emph{y}, and \emph{detIndices} (in that order). When such consolidated 
representation of the detection data x is used, \emph{detIndices} and \emph{detNums} arguments should not be specified.}

\item{Id.vector}{A matrix giving the x- and y-coordinates of each trap (scaled to the habitat; see (\code{\link{scaleCoordsToHabitatGrid}}).}

\item{idAugmented}{A matrix giving the x- and y-coordinates of each trap (scaled to the habitat; see (\code{\link{scaleCoordsToHabitatGrid}}).}

\item{lowerCoords, upperCoords}{Matrices of lower and upper x- and y-coordinates of all habitat windows scaled to the habitat (see (\code{\link{scaleCoordsToHabitatGrid}}).
One row for each window. Each window should be of size 1x1.}

\item{habitatGrid}{Matrix of habitat window indices. Cell values should correspond to the order of habitat windows in \code{lowerCoords} and \code{upperCoords}. 
When the habitat grid only consists of a single row or column of windows, an additional row or column of dummy indices has to be added because the \code{nimble} model code requires a matrix.}

\item{sd}{Standard deviation of the isotropic bivariate normal distribution.}

\item{movementMethod}{a character denoting the nimbleSCR to be used.}
}
\value{
This function returns a 2- or 3-dimensional array of initial AC locations.
}
\description{
R utility function to generate initial AC locations using compiled Nimble functions for efficiency.
}
\details{
The \code{getSInits} function is used in advance of model building.
}
\author{
Pierre Dupont
}
