% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rovquantR_data.R
\docType{data}
\name{GLOBALMAP}
\alias{GLOBALMAP}
\title{GLOBALMAP}
\format{
## `GLOBALMAP`
A Simple feature collection with 239 features and 4 fields
}
\source{
<https://www.nmbu.no/en/research/projects/rovquant>
}
\usage{
GLOBALMAP
}
\description{
These data sets describe Scandinavian countries, counties and management regions.

`COUNTIES`, `COUNTRIES`, `COUNTRIESWaterHumans` and `GLOBALMAP` are simple feature (sf) objects.
`habitatRasterResolution` and `habitatRasters` are raster objects.
}
\examples{
data(GLOBALMAP)
}
\keyword{datasets}
