% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rovquantR_data.R
\docType{data}
\name{habitatRasterResolution}
\alias{habitatRasterResolution}
\title{habitatRasterResolution}
\format{
## `habitatRasterResolution`
A list of RasterStacks
}
\source{
<https://www.nmbu.no/en/research/projects/rovquant>
}
\usage{
habitatRasterResolution
}
\description{
These data sets describe Scandinavian countries, counties and management regions.

`COUNTIES`, `REGIONS`, `COUNTRIES`, `COUNTRIESWaterHumans` and `GLOBALMAP` are simple feature (sf) objects.
`habitatRasterResolution` and `habitatRasters` are raster objects.
}
\examples{
data(habitatRasterResolution)
}
\keyword{datasets}
