% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rovquantR_data.R
\docType{data}
\name{habitatRasters}
\alias{habitatRasters}
\title{habitatRasters}
\format{
## `habitatRasters`
A RasterStack
}
\source{
<https://www.nmbu.no/en/research/projects/rovquant>
}
\usage{
habitatRasters
}
\description{
These data sets describe Scandinavian countries, counties and management regions.

`COUNTIES`, `REGIONS`, `COUNTRIES`, `COUNTRIESWaterHumans` and `GLOBALMAP` are simple feature (sf) objects.
`habitatRasterResolution` and `habitatRasters` are raster objects.
}
\examples{
data(habitatRasters)
}
\keyword{datasets}
