% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeAugmentation.R, R/MakeZfromScratch.R
\name{MakeAugmentation}
\alias{MakeAugmentation}
\alias{MakeZfromScratch}
\title{MakeAugmentation function}
\usage{
MakeAugmentation(y, aug.factor = NULL, aug.years = NULL, replace.value = NA)

MakeZfromScratch(data.alive, data.dead = NULL, samplingMonths = 1:12)
}
\arguments{
\item{y}{A \code{Vector}, \code{Matrix} or \code{Array} object containing the individual detection histories.}

\item{aug.factor}{A \code{numeric} object defining the augmentation factor to be used.}

\item{aug.years}{A \code{numeric} object defining the number of years to be added to the data (e.g. for population forecasting).}

\item{replace.value}{A \code{numeric} object defining the value to be repeated for augmented individuals.}

\item{data.alive}{A \code{sf dataframe} object containing individual detections.}
}
\value{
A \code{Vector}, \code{Matrix} or \code{Array} object containing the augmented y.

A \code{Vector}, \code{Matrix} or \code{Array} object containing the augmented y.
}
\description{
\code{MakeAugmentation} increases the dimensions of an object along
 the individual and/or the year dimension. It returns a \code{Vector}, \code{Matrix} 
 or \code{Array} object with the expanded object.

\code{MakeZfromScratch} creates a sequence of states for each individual
that is compatile with its detection history, i.e. "unborn" before the first detection,
"alive" between the first and last detection and "dead" after that.
 the dimensions of an object along
}
\examples{
\dontrun{MakeAugmentation()}

\dontrun{MakeZfromScratch()}

}
