% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeDetectionHistory.R
\name{makeDetectionHistory}
\alias{makeDetectionHistory}
\title{Detection history formatting}
\usage{
makeDetectionHistory(
  data,
  detectors,
  detection.fun,
  all.ids = NULL,
  diff.occ = FALSE,
  M = NULL
)
}
\arguments{
\item{data}{a data.frame (or list of data.frames if the model integrates multiple datasets) containing individual detections (in rows).
Mandatory column names are "id" and "detector",
Optional column names are "x", "y", "session" and "occasion".}

\item{detectors}{a data.frame (or list of data.frames if the model integrates multiple datasets) containing the coordinates of all possible detection locations.}

\item{detection.fun}{a character string indicating the nimbleSCR (or nimble) function used to model individual detections.
Must be one of :
\itemize{
\item "dbern", "dbinom", "dpois": Bernoulli, Binomial or Poisson distributions to model binary detections or detection frequencies on detection location at a time.
\item "dbinomLocal_normal", "dpoisLocal_normal": Local-evaluation functions; used to model a vector of individual detections at multiple discrete detectors. 
\item "dbernppDetection_normal", "dpoisppDetection_normal": Bernoulli and Poisson point-process functions; used to model SCR detections in continuous space.
\item "dbernppLocalDetection_normal", "dpoisppLocalDetection_normal": local evaluation versions of the two funcctions above for faster implementation.
}}

\item{all.ids}{vector of individual identifiers to be included in the detection history array. Can be used to either}

\item{diff.occ}{An aggregation factor to reduce the number of habitat cells to 
retrieve local objects for. Defaults to 1; no aggregation.}

\item{M}{Augmented population size.}
}
\value{
This function returns a detection history array in one of the following 
format:
\itemize{
\item y[]: ...
\item y[]: ...
\item y[]: ...
\item y[]: ...
}
}
\description{
R utility function to format SCR and OPSCR detection data into detection history arrays of different formats.
The format of the detection history is determined by the name of the \code{nimble} distribution used in the model (provided by the 'detection.fun' argument)

The \code{makeDetectionHistory} function is used in the data2nimbleSCR function to help
formatting SCR data for use with nimbleSCR.
}
\examples{
n.samples <- 150
n.ids <- 40
n.detectors <- 10
n.sessions <- 3
n.occasions <- 5

detectors <- expand.grid(list("id" = 1:n.detectors,
                              #"occasion" = 1:n.occasions,
                              "session"=1:n.sessions))
detectors <- cbind.data.frame("id" = sample(n.detectors,size = n.samples,replace = TRUE),
                              "session" = sample(n.sessions,size = n.samples,replace = TRUE),
                              "occasion" = sample(n.occasions,size = n.samples,replace = TRUE))
head(detectors)

data <- cbind.data.frame("id" = sample(n.ids,size = n.samples,replace = TRUE),
                         "detector" = sample(n.detectors,size = n.samples,replace = TRUE),
                         "session" = sample(n.sessions,size = n.samples,replace = TRUE),
                         "occasion" = sample(n.occasions,size = n.samples,replace = TRUE))
head(data)

all.ids <- 1:n.ids

makeDetectionHistory(data, detectors, "dbern", all.ids)
makeDetectionHistory(data, detectors, "dbinom", all.ids)
makeDetectionHistory(data, detectors, "dbinom_vector", all.ids)
makeDetectionHistory(data, detectors, "dpois", all.ids)
makeDetectionHistory(data, detectors, "dcat", all.ids)
makeDetectionHistory(data, detectors, "dbinomLocal_normal", all.ids)
makeDetectionHistory(data, detectors, "dpoisLocal_normal", all.ids)
makeDetectionHistory(data, detectors, "dbernppLocalDetection_normal", all.ids)
makeDetectionHistory(data, detectors, "dpoisppLocalDetection_normal", all.ids)

}
\author{
Pierre Dupont
}
