% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeDirectories.R
\name{makeDirectories}
\alias{makeDirectories}
\title{Prepare folder structure}
\usage{
makeDirectories(path = NULL, twoSex = T)
}
\arguments{
\item{path}{A path indicating where to create the different folders.}

\item{twoSex}{A \code{logical}. If TRUE, sex-specific folders are created for the female and male inputs and outputs files, respectively.}
}
\value{
boolean invisible(FALSE) if nothing was created, invisible(TRUE) if folders were created in \emph{path}.
}
\description{
\code{makeDirectories} creates a set of folders to store all the necessary
data for hte OPSCR ROvQuant analysis.
}
\examples{
\dontrun{makeDirectories()}

}
