% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeHabitatFromRastersf.R
\name{MakeHabitatFromRastersf}
\alias{MakeHabitatFromRastersf}
\title{Create a list of objects to define the habitat from a predefined habitat raster}
\usage{
MakeHabitatFromRastersf(
  poly,
  habitat.r,
  buffer = NULL,
  plot.check = TRUE,
  tolerance = NULL
)
}
\arguments{
\item{poly}{A \code{SpatialPolygon} object with the study area}

\item{habitat.r}{A \code{raster} defining suitable habitat.}

\item{buffer}{A \code{Numeric} with the size of the buffer area around the study area (in meters in poly is UTM).}

\item{plot.check}{A \code{Logical} to display checking plots (if TRUE).}

\item{tolerance}{A \code{Numeric}. Default is (res(habitat.r)[1]/5). 
This tolerance threshold is used to identify cells that should be considered in the buffer and not searched. 
Sometimes raster cells will be identified as not-searched but should actually be searched.
This is because the center of the cell does not fall in the polygon used to create the searched area.
Instead we identify cells that are in the habitat but "far" from the poly.
Those cells are then considered as buffer.}
}
\value{
A \code{List} object with the coordinates and attributes of the habitat
 @return habitat.sp: A \code{SpatialPointsDataFrame} object with the coordinates of the center of cells of the habitat
 @return habitat.clip.sp: A \code{SpatialPointsDataFrame} object with the coordinates of the center of cells of the habitat clipped to the suitable habitat.
 @return habitat.xy: A \code{DataFrame} object with the x and y coordinates of the habitat cells.
 @return IDCells.mx: A \code{matrix} with the ID of each cell of the raster
 @return habitat.r: A \code{raster} with the suitable habitat (1) and non suitable (0)
 @return habitat.mx: A \code{matrix} with the suitable habitat (1) and non suitable (0)
 @return resolution: A \code{numeric} with the resolution used for the raster 
 @return buffered.habitat.poly: A \code{SpatialPolygons} that includes the buffer. 
 @return habitat.poly: A \code{SpatialPolygons} with the original polygon used. 
 @return habitat.index: A \code{vector} with ID cell of the raster that fall within the suitable habitat.
}
\description{
\code{MakeHabitatFromRaster} returns a \code{List} of objects necessary to define the habitat in SCR (can be slow for large habitat...)
}
