% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeRovquantData.R
\name{makeRovquantData}
\alias{makeRovquantData}
\title{RovQuant OPSCR data preparation.}
\usage{
makeRovquantData(
  data.dir = "./Data",
  working.dir = NULL,
  species = c("bear", "wolf", "wolverine"),
  years = NULL,
  sex = c("female", "male"),
  aug.factor = NULL,
  sampling.months = NULL,
  habitat.res = NULL,
  x.extent = NULL,
  y.extent = NULL,
  buffer.size = NULL,
  max.move.dist = NULL,
  detector.res = NULL,
  subdetector.res = NULL,
  max.det.dist = NULL,
  resize.factor = NULL,
  rename.list = NULL,
  print.report = TRUE,
  Rmd.template = NULL,
  output.dir = NULL,
  overwrite = FALSE
)
}
\arguments{
\item{data.dir}{A \code{path} to the directory containing the clean Rovbase data,
as prepared by \code{cleanRovBaseData}.}

\item{working.dir}{A \code{path} to the directory for this analysis containing
the \code{nimbleInFiles} folder to store the prepared data.}

\item{species}{A \code{character} denoting the species; can be any of "bear", "wolf" or "wolverine.}

\item{years}{A \code{numeric} vector denoting which years of data will be used for this analysis.}

\item{sex}{A \code{character} denoting for which sex the data should be prepared ("female", "male" or both).}

\item{aug.factor}{A \code{numeric} denoting the augmentation factor.}

\item{sampling.months}{A \code{list} denoting the months of the monitoring period. Can accomodate sampling periods over two calendar years, e.g. 'list(c(10:12),c(1:3))' for a monitoring period extending from October to March (included).}

\item{habitat.res}{A \code{numeric} denoting the resolution (in meters) to be used for the habitat definition.}

\item{buffer.size}{A \code{numeric} denoting the size (in meters) of the buffer around the searched area.}

\item{max.move.dist}{A \code{numeric} denoting the maximum allowed distance for inter-annual movement.}

\item{detector.res}{A \code{numeric} denoting the resolution (in meters) to be used for the detectors definition.}

\item{subdetector.res}{A \code{numeric} denoting the resolution (in meters) to be used for the sub-detectors.}

\item{max.det.dist}{A \code{numeric} denoting the maximum allowed distance for intra-annual movement.}

\item{resize.factor}{A \code{numeric} denoting the aggregation factor to be used for the local evaluation.}

\item{print.report}{A \code{logical} denoting whether to print out a \code{.html} report summarizing the data preparation process or not.}

\item{Rmd.template}{(optional) A \code{path} to a custom .Rmd template to use instead of the default one provided in 'rovquantR'.}

\item{output.dir}{(optional) A \code{path} denoting where to print the .html report. Default is in a folder named 'reports' in the working directory.}

\item{overwrite}{A \code{logical} (default = FALSE) to force overwriting of previously existing data.
If FALSE, the function checks for any pre-existing data files and asks whether to overwrite it or not.}
}
\value{
This function returns:
\enumerate{
\item Multiple \code{.RData} nimble input files containing the input data, model code and initial values necessary for fitting the model in nimble (one input file per MCMC chain).
\item A \code{.html} report summarizing the data preparation process. 
\item Additional \code{.png} images and \code{.csv} tables of the figures and tables present in the report.
}
}
\description{
The \code{makeRovquantData} function calls a custom Rmarkdown template that identifies 
and loads the most recent Rovbase data available for the specified species 
and performs the OPSCR data preparation for model fitting. 

The data preparation process is composed of three main steps:
\enumerate{
\item Defining and formatting habitat characteristics
\item Defining and formatting detectors characteristics
\item Defining and formatting individual detection histories
}

The data preparation process is specific to each species and incorporates the
different developments from project RovQuant to allow fitting large-scale
SCR and OPSCR models (see Turek et al., 2021 <doi.org/10.1002/ecs2.3385>  for more details):
\enumerate{
\item Rescaling spatial coordinates to a grid for fast look-up assignment.
\item A local evaluation approach (see Milleret et al., 2019 <doi:10.1002/ece3.4751> for more details)
\item A sparse matrix representation of the observation data to reduce the size of objects to be processed.
\item An indicator to shortcut calculations for individuals unavailable for detection.
}
}
\author{
Pierre Dupont
}
