% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeRovquantData_wolf.R,
%   R/makeRovquantData_wolf_OLD.R
\name{makeRovquantData_wolf}
\alias{makeRovquantData_wolf}
\alias{makeRovquantData_wolf_OLD}
\title{RovQuant OPSCR wolf data preparation.}
\usage{
makeRovquantData_wolf(
  data.dir = getwd(),
  working.dir = getwd(),
  years = NULL,
  sex = c("female", "male"),
  aug.factor = 0.8,
  sampling.months = list(10:12, 1:3),
  habitat.res = 20000,
  x.extent = NULL,
  y.extent = NULL,
  buffer.size = 40000,
  max.move.dist = 250000,
  detector.res = 10000,
  subdetector.res = 1000,
  max.det.dist = 45000,
  resize.factor = 1,
  rename.list = NULL
)

makeRovquantData_wolf_OLD(
  data.dir = getwd(),
  working.dir = getwd(),
  years = NULL,
  sex = c("female", "male"),
  aug.factor = 0.8,
  sampling.months = list(10:12, 1:3),
  habitat.res = 20000,
  x.extent = NULL,
  y.extent = NULL,
  buffer.size = 40000,
  max.move.dist = 250000,
  detector.res = 10000,
  subdetector.res = 1000,
  max.det.dist = 45000,
  resize.factor = 1,
  rename.list = NULL
)
}
\arguments{
\item{data.dir}{A \code{path}.}

\item{working.dir}{A \code{path}.}

\item{years}{A \code{list}.}

\item{sex}{A \code{character}.}

\item{aug.factor}{A \code{Numeric}.}

\item{sampling.months}{A \code{list}.}

\item{habitat.res}{A \code{Numeric}.}

\item{x.extent}{(Optional) A \code{Numeric}.}

\item{y.extent}{(Optional) A \code{Numeric}.}

\item{buffer.size}{A \code{Numeric}.}

\item{max.move.dist}{A \code{Numeric}.}

\item{detector.res}{A \code{Numeric}.}

\item{subdetector.res}{A \code{Numeric}.}

\item{max.det.dist}{A \code{Numeric}.}

\item{resize.factor}{A \code{Numeric}.}

\item{rename.list}{A \code{list}.}
}
\value{
A \code{html} report summarizing the data preparation process
Additional \code{.png} images that can be reused somewhere else.

A \code{html} report summarizing the data preparation process
Additional \code{.png} images that can be reused somewhere else.
}
\description{
\code{makeRovquantData_wolf} formats the available wolf data for the OPSCR analysis using nimble and nimbleSCR.
The data preparation process is composed of three main steps:
 - defining and formatting habitat characteristics
 - defining and formatting detectors characteristics
 - defining and formatting individual detection histories

\code{makeRovquantData_wolf} formats the available wolf data for the OPSCR analysis using nimble and nimbleSCR.
The data preparation process is composed of three main steps:
 - defining and formatting habitat characteristics
 - defining and formatting detectors characteristics
 - defining and formatting individual detection histories
}
\author{
Pierre Dupont
}
