% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeRovquantData_wolverine.R
\name{makeRovquantData_wolverine}
\alias{makeRovquantData_wolverine}
\title{RovQuant OPSCR wolverine data preparation.}
\usage{
makeRovquantData_wolverine(
  data.dir = getwd(),
  working.dir = getwd(),
  years = NULL,
  sex = c("female", "male"),
  aug.factor = 0.8,
  sampling.months = list(12, 1:6),
  habitat.res = 20000,
  buffer.size = 60000,
  max.move.dist = 250000,
  detector.res = 10000,
  subdetector.res = 2000,
  max.det.dist = 40000,
  resize.factor = 1,
  rename.list = NULL
)
}
\arguments{
\item{data.dir}{A \code{path}.}

\item{working.dir}{A \code{path}.}

\item{years}{A \code{list}.}

\item{sex}{A \code{character}.}

\item{aug.factor}{A \code{Numeric}.}

\item{sampling.months}{A \code{list}.}

\item{habitat.res}{A \code{Numeric}.}

\item{buffer.size}{A \code{Numeric}.}

\item{max.move.dist}{A \code{Numeric}.}

\item{detector.res}{A \code{Numeric}.}

\item{subdetector.res}{A \code{Numeric}.}

\item{max.det.dist}{A \code{Numeric}.}

\item{resize.factor}{A \code{Numeric}.}

\item{rename.list}{A \code{list}.}
}
\value{
A \code{html} report summarizing the data preparation process
Additional \code{.png} images that can be reused somewhere else.
}
\description{
\code{makeRovquantData_wolverine} formats the available wolverine data for the OPSCR analysis using nimble and nimbleSCR.
The data preparation process is composed of three main steps:
 - defining and formatting habitat characteristics
 - defining and formatting detectors characteristics
 - defining and formatting individual detection histories
}
\author{
Pierre Dupont
}
