% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeSearchGrid.R
\name{makeSearchGrid}
\alias{makeSearchGrid}
\title{Function to create a detector grid}
\usage{
makeSearchGrid(data, resolution, div = 1, center = T, plot = TRUE)
}
\arguments{
\item{data}{A \code{sf} object. can be a \code{SpatialPolygons} or \code{SpatialPoints} or \code{SpatialLines} or \code{raster}. Spatial Lines objects takes time to compute}

\item{resolution}{Numeric variable denoting the size of grid cells in units of (\code{polygon}).}

\item{div}{Numeric variable denoting the number of equally-sized subplots grid cell is to be divided into.}

\item{center}{Logical variable denoting whether nodes of the resulting search grid are to be centered within their respective subplots.}

\item{plot}{Logical for whether (\code{TRUE}) or not (\code{FALSE}) check plots are to be generated.}
}
\description{
\code{makeSearchGrid} creates equally spaced detectors within a defined area (data) using 
a given resolution (i.e. distance between two detectors).
If specified, each detector grid can be divived into multiple equally spaced sub-detectors (e.g. for use in the PAB model).
}
\author{
Richard Bischof, \email{richard.bischof@nmbu.no} and Pierre Dupont
}
