% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeTrapResponseCov.R
\name{makeTrapResponseCov}
\alias{makeTrapResponseCov}
\title{Create a trap response covariate}
\usage{
makeTrapResponseCov(data, data.dead = NULL, IDs = NULL)
}
\arguments{
\item{data}{A \code{sf} object containing the detection data}

\item{data.dead}{OPTIONAL: A second \code{sf} object containing the dead recovery data}

\item{IDs}{OPTIONAL: A \code{vector} of IDs to return the trap response covariate for.
Any detection further than \code{radius}m from any detector will not be assigned to any detector. Instead it will get a 'NA'.}
}
\value{
A \code{matrix} object with the x and y locations of the
 different detections with associated detector's Id
}
\description{
\code{makeTrapResponseCov} ide and assigns the closest detector to each detection in \code{data}.
}
