% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeY.R
\name{makeY}
\alias{makeY}
\title{Generate Detection Histories Arrays}
\usage{
makeY(
  data,
  detectors,
  method = "Bernoulli",
  data2 = NULL,
  detectors2 = NULL,
  method2 = "Bernoulli",
  IDs = NULL,
  returnIdvector = FALSE
)
}
\arguments{
\item{data}{A \code{SpatialPointsDataFrame} of detections with ID, Year and Detector columns.}

\item{detectors}{A \code{list} of SpatialPointsDataFrames containing the detector locations.}

\item{method}{A \code{Character} object defining the array to be returned ("Binomial" , "Poisson" , "Bernoulli").}

\item{data2}{OPTIONAL: A second \code{SpatialPointsDataFrame} of detections with ID, Year and Detector columns.}

\item{detectors2}{OPTIONAL: A \code{SpatialPointsDataFrame}of SpatialPointsDataFrames containing the detector locations.}

\item{method2}{OPTIONAL: A \code{Character} object defining the array to be returned ("Binomial" , "Poisson" , "Bernoulli").}

\item{IDs}{OPTIONAL: A \code{vector} of IDs to return the y array for.}

\item{returnIdvector}{OPTIONAL: A \code{logical} to return the vector of individual ID.}
}
\value{
A \code{array} object containing the number of detections per individual (rows), detector (columns), and years (third dimensions).
}
\description{
\code{makeY} returns a \code{array} object with the detections (presence/absence or counts) of individuals over detectors and time.
}
\author{
Pierre Dupont
}
