% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeY.R
\name{MakeYsf}
\alias{MakeYsf}
\title{Generate Detection Histories Arrays}
\usage{
MakeYsf(
  myData,
  myDetectors,
  method = "Bernoulli",
  myData2 = NULL,
  myDetectors2 = NULL,
  method2 = "Bernoulli",
  Id = NULL,
  returnIdvector = FALSE
)
}
\arguments{
\item{myData}{A \code{SpatialPointsDataFrame} of detections with ID, Year and Detector columns.}

\item{myDetectors}{A \code{list} of SpatialPointsDataFrames containing the detector locations.}

\item{method}{A \code{Character} object defining the array to be returned ("Binomial" , "Poisson" , "Bernoulli").}

\item{myData2}{OPTIONAL: A second \code{SpatialPointsDataFrame} of detections with ID, Year and Detector columns.}

\item{myDetectors2}{OPTIONAL: A \code{SpatialPointsDataFrame}of SpatialPointsDataFrames containing the detector locations.}

\item{method2}{OPTIONAL: A \code{Character} object defining the array to be returned ("Binomial" , "Poisson" , "Bernoulli").}

\item{Id}{OPTIONAL: A \code{vector} of IDs to return the y array for.}

\item{returnIdvector}{OPTIONAL: A \code{logical} to return the vector of individual ID.}
}
\value{
A \code{Array} object with the individual in rows, the detectors in columns and years in third dimension.
}
\description{
\code{MakeY} returns a \code{Array} object with the detections (presence/absence or counts) of individuals over detectors and time.
}
\author{
Pierre Dupont
}
