% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotQuantiles.R
\name{plotQuantiles}
\alias{plotQuantiles}
\title{Plot quantiles}
\usage{
plotQuantiles(
  x,
  at,
  width = 0.15,
  quantiles = c(0.0275, 0.975),
  quantiles2 = c(0.25, 0.75),
  col = adjustcolor("firebrick3", 0.5),
  add.star = FALSE
)
}
\arguments{
\item{x}{Samples for which to plot the quantiles.}

\item{at}{x-axis coordinates of the bar.}

\item{width}{width of the bar.}

\item{quantiles}{a vector of two numeric values denoting the quantiles to be represented.}

\item{quantiles2}{a second vector of numeric values denoting the second set of quantiles to be represented.}

\item{col}{color to be used for plotting.}
}
\value{
This function plots quantile bars with the specified characteristics.
}
\description{
R utility function to plot the quantiles of a distribution of samples as a bar.
}
\examples{
  
plot(1, xlim = c(0, 2), ylim = c(0,100), type = "n")
plotQuantiles(x = rnorm(1000, 60, sd = 12), at = 1, width = 0.18, quantiles2 = c(0.25,0.75))

}
\author{
Pierre Dupont
}
