% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/populate.R
\name{populate}
\alias{populate}
\alias{get.dim}
\title{Populate coda output}
\usage{
get.dim(params)

populate(input, dim, simslist = FALSE, samples = NULL)
}
\arguments{
\item{params}{A \code{character} object.}

\item{input}{A \code{logical} if DIC statistics should be computed or not.}

\item{dim}{A \code{vector} of parameters for which statistics should not be computed.}

\item{simslist}{A \code{logical}. If TRUE, messages are displayed.}

\item{samples}{A \code{logical}. If TRUE, messages are displayed.}
}
\value{
A \code{list} with sims lists, parameter summary values and rhat values.
}
\description{
\code{populate} and \code{get.dim} are internal functions used in \code{ProcessCodaOutput}.
}
