% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processRovquantOutput.R
\name{processRovquantOutput}
\alias{processRovquantOutput}
\title{RovQuant OPSCR output processing}
\usage{
processRovquantOutput(
  data.dir = "./Data",
  working.dir = NULL,
  species = c("bear", "wolf", "wolverine"),
  nburnin = 5,
  niter = 100,
  extraction.res = 5000,
  print.report = TRUE,
  Rmd.template = NULL,
  output.dir = NULL,
  overwrite = FALSE,
  full.report = FALSE
)
}
\arguments{
\item{data.dir}{A \code{path} to the directory containing the clean Rovbase data, as prepared by \code{cleanRovBaseData}.}

\item{working.dir}{A \code{path} to the directory for this analysis containing the \code{nimbleInputFiles} folder to store the prepared data.}

\item{species}{A \code{character} denoting the species; can be one of "bear", "wolf" or "wolverine.}

\item{nburnin}{An \code{integer} denoting the number of MCMC bites to be removed from each MCMC chain as burnin.}

\item{niter}{An \code{integer} denoting the number of MCMC iterations to be used for density extraction.}

\item{extraction.res}{A \code{numeric} denoting the resolution (in meters) to use for density extraction.}

\item{print.report}{A \code{logical}  Whether an .html report summarizingf the results should be printed (TRUE) or not (FALSE).}

\item{Rmd.template}{(optional) A \code{path} to a custom .Rmd template to use instead of the default one provided in 'rovquantR'.}

\item{output.dir}{(optional) A \code{path} where to print the report. Default is to print in the 'reports' folder of the working directory.}

\item{overwrite}{A \code{logical} Whether to overwrite (TRUE) or ask before overwriting potentially existing .html reports (FALSE).}
}
\value{
This function returns:
\enumerate{
\item Multiple \code{.RData} files with the processed MCMC outputs and density outputs.
\item A \code{.html} report summarizing the data cleaning process. 
\item Additional \code{.png} images and \code{.csv} that can be reused somewhere else.
}
}
\description{
The \code{processRovquantOutput} function calls a custom Rmarkdown template that identifies 
and loads the most recent Rovbase data available for the specified species 
and performs the OPSCR data preparation for model fitting. 
The data preparation is specific to each species and incorporates the
different developments from project RovQuant to allow fitting large-scale
SCR and OPSCR models (see Turek et al., 2021 <doi.org/10.1002/ecs2.3385>  for more details):
\enumerate{
\item Rescaling spatial coordinates to a grid for fast look-up assignment.
\item A local evaluation approach (see Milleret et al., 2019 <doi:10.1002/ece3.4751> for more details)
\item A sparse matrix representation of the observation data to reduce the size of objects to be processed.
\item An indicator to shortcut calculations for individuals unavailable for detection.
}
}
\author{
Pierre Dupont
}
