% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processRovquantOutput_bear.R
\name{processRovquantOutput_bear}
\alias{processRovquantOutput_bear}
\title{RovQuant OPSCR bear output processing}
\usage{
processRovquantOutput_bear(
  data.dir = getwd(),
  working.dir = NULL,
  nburnin = 0,
  niter = 100,
  extraction.res = 5000,
  overwrite = FALSE
)
}
\arguments{
\item{data.dir}{A \code{path} to the directory containing the clean Rovbase data, as prepared by \code{cleanRovBaseData}.}

\item{working.dir}{A \code{path} to the directory for this analysis containing the \code{nimbleInputFiles} folder to store the prepared data.}

\item{nburnin}{An \code{integer} denoting the number of MCMC bites to be removed from each MCMC chain as burnin.}

\item{niter}{An \code{integer} denoting the number of MCMC iterations to be used for density extraction.}

\item{extraction.res}{A \code{integer} denoting the raster resolution for density extraction.}

\item{overwrite}{A \code{logical} Whether to overwrite (TRUE) or ask before overwriting potentially existing output files (FALSE).}
}
\value{
Multiple \code{.RData} files with the processed MCMC outputs and density outputs.
Additional \code{.png} images and \code{.csv} that can be reused somewhere else.
}
\description{
\code{processRovquantOutput_bear} calls a custom Rmarkdown template that combines 
and processes MCMC outputs from NIMBLE models and produces figures,
tables and rasters of interest (e.g. population density maps)
}
\author{
Pierre Dupont
}
