% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processRovquantOutput_wolf.R
\name{processRovquantOutput_wolf}
\alias{processRovquantOutput_wolf}
\title{RovQuant OPSCR wolf output processing}
\usage{
processRovquantOutput_wolf(
  data.dir = getwd(),
  working.dir = NULL,
  nburnin = 0,
  niter = 100,
  thin = 1,
  thin2 = 1,
  extraction.res = 5000,
  overwrite = FALSE
)
}
\arguments{
\item{data.dir}{A \code{path}}

\item{working.dir}{A \code{path}}

\item{nburnin}{An \code{integer} denoting the number of iterations to be removed from each MCMC as burnin.}

\item{niter}{An \code{integer} denoting the number of MCMC iterations to be used for density extraction.}

\item{thin}{An \code{integer} denoting the thinning rate for main parameters.}

\item{thin2}{An \code{integer} denoting the thinning rate for secondary parameters.}

\item{extraction.res}{A \code{integer} denoting the raster resolution for density extraction.}

\item{overwrite}{A \code{logical} denoting whether to overwrite existing.}
}
\value{
\code{.RData} files with the MCMC samples, summary statistics and extracted population densities and population sizes.
A \code{html} report summarizing the main results from the OPSCR analysis.
Additional \code{.png} images that can be reused somewhere else.
}
\description{
\code{processRovquantOutput_wolf} calls a custom Rmarkdown template that combines 
and processes MCMC outputs from NIMBLE models and produces figures,
tables and rasters of interest (e.g. population density maps)
}
\author{
Pierre Dupont
}
