% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{quantileCpp}
\alias{quantileCpp}
\title{Quantile calculation}
\usage{
quantileCpp(x, q)
}
\arguments{
\item{x}{A numeric vector whose sample quantiles are wanted.}

\item{q}{A numeric vector of probabilities.}
}
\description{
This function returns sample quantiles corresponding to the given 
probabilities (\code{q}). The smallest observation corresponds to a 
probability of 0 and the largest to a probability of 1.
}
