% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readMostRecent.R
\name{readMostRecent}
\alias{readMostRecent}
\alias{readMostRecent.csv}
\alias{readMostRecent.excel}
\alias{readMostRecent.RData}
\alias{getMostRecent}
\alias{writeMostRecent.csv}
\title{Read most recent version of a file}
\usage{
readMostRecent(
  path,
  extension = ".csv",
  pattern = NULL,
  returnDate = FALSE,
  sep = ",",
  dec = ".",
  stack = TRUE,
  ...
)

readMostRecent.csv(path, returnDate = F, fileEncoding = NULL, ...)

readMostRecent.excel(path, returnDate = F, ...)

readMostRecent.RData(path, pattern = ".RData", returnDate = F, ...)

getMostRecent(path, pattern)

writeMostRecent.csv(x, file)
}
\arguments{
\item{path}{A character string denoting the data directory path.}

\item{extension}{A character string denoting the type of file to be loaded (can be one of ".csv", ".RData", "xls", "xlsx")}

\item{pattern}{(Optional) An additional character string to be matched with the file name.}

\item{returnDate}{Logical. Whether to return the date the loaded file was last modified.}

\item{sep}{The field separator character used when loading \code{".csv"} files. Values on each line of the file are separated by this character (default is ","),.}

\item{dec}{The character used in the file for decimal points.}

\item{...}{additional optional parameters.}

\item{fileEncoding}{Character string: if non-empty declares the encoding used on a file when given as a character string (see \code{\link[utils]{read.csv}} for more details).}

\item{x}{The object to be written, preferably a matrix or data frame. If not, it is attempted to coerce x to a data frame.}

\item{file}{Either a character string naming a file or a connection open for writing. "" indicates output to the console.}
}
\value{
The data loaded
}
\description{
The \code{readMostRecent} functions provide a suite of utility functions to 
identify and load the most recent file matching a set of characteristics.
}
\author{
Pierre Dupont
}
