% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readMultiples.R
\name{readMultiples}
\alias{readMultiples}
\title{Read and combine multiple files}
\usage{
readMultiples(
  path,
  extension = ".csv",
  pattern = NULL,
  sep = ",",
  dec = ".",
  ...
)
}
\arguments{
\item{path}{A character string denoting the data directory path.}

\item{extension}{A character string denoting the type of file to be loaded (can be one of ".csv", ".RData", "xls", "xlsx")}

\item{pattern}{(Optional) An additional character string to be matched with the file name.}

\item{sep}{The field separator character used when loading \code{".csv"} files. Values on each line of the file are separated by this character (default is ","),.}

\item{dec}{The character used in the file for decimal points.}

\item{...}{additional optional parameters.}
}
\value{
The data loaded
}
\description{
The \code{readMultiples} function identifies and loads multiple files matching a set of characteristics, before combining them using rbind.
}
\author{
Pierre Dupont
}
