% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readTracks.R
\name{readTracks}
\alias{readTracks}
\title{Read and combine multiple .shp files containing GPS tracks of recorded search effort}
\usage{
readTracks(
  data.dir,
  years = 1950:format(Sys.Date(), "\%Y"),
  sampling.months = 1:12
)
}
\arguments{
\item{data.dir}{A character string denoting the data directory path.}
}
\value{
The data loaded
}
\description{
The \code{readTracks} function identifies and loads multiple .shp files, before combining them using rbind.
}
\author{
Pierre Dupont
}
