% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanitize.R
\name{sanitize}
\alias{sanitize}
\alias{sanitizeInput}
\alias{fixNames}
\title{Sanitize nimbleSCR input data}
\usage{
sanitizeInput(data, detectors)

fixNames(data, name = NULL, is.data = TRUE)
}
\arguments{
\item{data}{a data.frame (or list of data.frames if the model integrates multiple datasets) containing individual detections (in rows).}

\item{name}{(optional) a character string with the name of the data frame tested (used to report potential errors).}
}
\description{
R utility function to sanitize one or multiple individual detection datasets together with the corresponding detectors information.
This function ensures data format and names are compatible with rovquantR and nimbleSCR functions.

The \code{sanitizeData} function is used among other in the data2nimbleSCR function to help formatting SCR and OPSCR data for use with nimbleSCR.
}
\examples{
n.samples <- 150
n.ids <- 40
n.detectors <- 10
n.sessions <- 3
n.occasions <- 5

## Example 1: single dataset - OPSCR
dataset1 <- cbind.data.frame( "iD" = sample(n.ids,size = n.samples,replace = TRUE),
                             "detECtor" = sample(n.detectors,size = n.samples,replace = TRUE),
                             "sub.detector" = sample(n.occasions, size = n.samples,replace = TRUE),
                             "ids.covs" = sample(n.occasions, size = n.samples,replace = TRUE),
                             "Year" = sample(n.sessions,size = n.samples,replace = TRUE),
                             "occasion" = sample(n.occasions,size = n.samples,replace = TRUE))
head(dataset1)

detectors <- expand.grid(list( "id" = 1:n.detectors,
                               #"occasion" = 1:n.occasions,
                               "session" = 1:n.sessions))
DETS <- cbind.data.frame( "detectors" = sample(n.detectors,size = n.samples,replace = TRUE),
                          "session" = sample(n.sessions,size = n.samples,replace = TRUE),
                          "occasion" = sample(n.occasions,size = n.samples,replace = TRUE))
head(DETS)

sanitizeInput(dataset1, DETS)
lapply(dataset1, head)
lapply(DETS, head)


## Example 2: multiple datasets - OPSCR
myData <- list( data_1 =  cbind.data.frame( "IDs" = sample(n.ids,size = n.samples,replace = TRUE),
                                            "detECtor" = sample(n.detectors,size = n.samples,replace = TRUE),
                                            "Years" = sample(n.sessions,size = n.samples,replace = TRUE),
                                            "Occasion" = sample(n.sessions,size = n.samples,replace = TRUE),
                                            "sub.detector" = sample(n.occasions,size = n.samples,replace = TRUE)),
                SCR = cbind.data.frame( "id" = sample(n.ids,size = n.samples,replace = TRUE),
                                        "DeteCtors" = sample(n.detectors,size = n.samples,replace = TRUE),
                                        "Occasion" = sample(n.sessions,size = n.samples,replace = TRUE)))
lapply(myData, head)

DETS <- list( cbind.data.frame( "detectors" = sample(n.detectors,size = n.samples,replace = TRUE),
                                "session" = sample(n.sessions,size = n.samples,replace = TRUE),
                                "occasion" = sample(n.occasions,size = n.samples,replace = TRUE),
                                "cov1" = rnorm(n.samples)),
              cbind.data.frame( "detectors" = sample(n.detectors,size = n.samples,replace = TRUE),
                                "occasion" = sample(n.occasions,size = n.samples,replace = TRUE)))
lapply(DETS,head)

sanitizeInput(myData,DETS)
lapply(myData, head)
lapply(DETS, head)
}
\references{
nimbleSCR manual: ...
}
\author{
Pierre Dupont
}
