% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/substrRight.R
\name{substrRight}
\alias{substrRight}
\title{Character subsetting}
\usage{
substrRight(x, n)
}
\arguments{
\item{x}{a \code{character} string to substract.}

\item{n}{a \code{integer} indicating the number of characters to return}
}
\value{
A vector of window sizes.
}
\description{
Identifies and returns the \code{n} last characters from the character string \code{x}.
}
\examples{
x <- "this.character.string"
n <- 6
substrRight(x, n)

}
\author{
Pierre Dupont
}
