% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeREADME.R
\name{writeREADME}
\alias{writeREADME}
\title{Write README file for the analysis}
\usage{
writeREADME(
  path = NULL,
  template = NULL,
  name = "rovquantR analysis",
  save_as = "README.Rmd",
  overwrite = FALSE
)
}
\arguments{
\item{path}{A \code{path} indicating where to create the README file.}

\item{template}{(Optional) A \code{path} indicating where to find the template of the README file. Default is to use the template provided in the \code{rovquantR} package.}

\item{A}{\code{logical}. (Optional; requires package  \emph{fs}) If TRUE, the directory structure created is displayed in a tree-like format.}
}
\value{
boolean invisible(FALSE) if nothing was created, invisible(TRUE) if the README file was created created in \emph{path}.
}
\description{
\code{writeREADME} creates and opens a README.rmd file to document the analysis.
}
\examples{
\dontrun{writeREADME()}

}
\author{
Pierre Dupont
}
