% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignDetectors.R
\name{assignDetectors}
\alias{assignDetectors}
\title{Detectors Assignment Function}
\usage{
assignDetectors(data, detectors, subDetectors = NULL, radius = 20000)
}
\arguments{
\item{data}{A \code{sf} object containing the detection data}

\item{detectors}{A \code{list} of sf objects containing the detector locations}

\item{subDetectors}{Optional; a \code{list} of sf objects containing the subdetectors locations (for PAB models only)}

\item{radius}{a \code{numeric} value denoting the maximum radius to consider for detector assignment. 
Any detection further than \code{radius}m from any detector will not be assigned to any detector. Instead it will get a 'NA'.}
}
\value{
A \code{sf} object with the x and y locations of the
 different detections with associated detector's Id
}
\description{
\code{assignDetectors} identifies and assigns the closest detector to each detection in \code{data}.
}
