% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignSearchTracks.R
\name{assignSearchTracks}
\alias{assignSearchTracks}
\title{Assign DNA samples to search tracks}
\usage{
assignSearchTracks(data, tracks, progress.bar = T)
}
\arguments{
\item{data}{A \code{sf} object containing the coordinates of the detections to be designated to tracks.}

\item{tracks}{A \code{sf} object containing the search tracks}

\item{progress.bar}{a \code{logical}, whether to display a progress bar or not.}
}
\value{
A \code{sf} object with the x and y locations of the different detections with associated tracks Id
}
\description{
the \code{assignSearchtTracks} function is a R function used to match RovBase
DNA samples to the corresponding GPS search track. Matching is based on date 
(exact match) and distance (within a given threshold).
}
\author{
Pierre Dupont
}
