% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignSearchTracks.R
\name{assignSearchtTracks}
\alias{assignSearchtTracks}
\title{Assign DNA samples to search tracks bite-size NIMBLE MCMC algorithms}
\usage{
assignSearchtTracks(
  data = myFilteredData.sp$alive,
  tracks = TRACKS_YEAR,
  dist = 750,
  progress.bar = T
)
}
\arguments{
\item{progress.bar}{a \code{logical value}. if TRUE, a separate progress bar is 
printed for each MCMC chain.}

\item{mcmc}{a \code{NIMBLE MCMC algorithm} object.}

\item{model}{a \code{NIMBLE model} object.}

\item{conf}{a \code{NIMBLE MCMC configuration} object.}

\item{bite.size}{an \code{integer} denoting the number of MCMC iterations in 
each MCMC bite.}

\item{bite.number}{an \code{integer} denoting the number of MCMC bites to run.}

\item{path}{a \code{character string} denoting the path where MCMC bite outputs are
to be saved as .RData files (when using \code{runMCMCbites} or \code{restartMCMCbites}) 
or looked for (when using \code{collectMCMCbites}).}

\item{save.rds}{(optional) a \code{logical value}. if TRUE, the state of the
nimble model is saved at the end of every MCMC bite. This is later used to restart 
the MCMC sampling process if necessary using \code{restartMCMCbites}.}

\item{path.rds}{(optional) a \code{path} to the folder containing the last MCMC model state file.}

\item{burnin}{an \code{integer} denoting the number of MCMC bites to be removed 
as burn-in.}

\item{pattern}{a \code{character string} denoting the name of the object containing
MCMC samples in each .R Data bite file.}

\item{param.omit}{a \code{character vector} denoting the names of parameters 
that are to be ignored when combining MCMC bites.}

\item{samplerDef}{a  \code{NIMBLE sampler definition} object.}

\item{modelState}{\code{NIMBLE model state} object.}

\item{mcmcState}{\code{NIMBLE MCMC state} object.}
}
\value{
\code{runMCMCbites} and \code{restartMCMCbites} run and save locally 
nimble MCMC posterior samples. \code{collectMCMCbites} combines multiple MCMC
 bites (and MCMC chains) and returns them in the commnonly used \code{mcmc.list}
 format from the \code{coda} package.
}
\description{
the \code{assignSearchtTracks} function is a R function used to match RovBase
DNA samples to the corresponding search track. Matching is based on date 
(exact match) and distance (within a given threshold).
samples on the fly.
}
\author{
Pierre Dupont
}
