% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CheckDistanceDetections.R
\name{CheckDistanceDetections}
\alias{CheckDistanceDetections}
\title{Check individual distances between all detections of an individual and the centroid of its detections}
\usage{
CheckDistanceDetections(
  y = y,
  detector.xy = detector.xy,
  max.distance = NULL,
  method = "pairwise",
  plot.check = TRUE
)
}
\arguments{
\item{y}{A \code{matrix} or \code{array} of detections}

\item{detector.xy}{A \code{matrix} object with xy coordinates of detectors.}

\item{max.distance}{A \code{numeric} distance (radius) used to identify  which detections are too far.}

\item{method}{A \code{character} describing the method to be used to select which detections to remove.}

\item{plot.check}{A \code{logical} plot distance histogram if \code{TRUE}.}
}
\value{
A \code{list} object with the distances between detections and centroid of detections and which detection(s) from which id are outside the max.distance.
}
\description{
\code{CheckDistanceDetections} returns a \code{list} object with the distances between detections and centroid of detections and which detection from which id is outside the max.distance if specified
}
