% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkInputValidity.R
\name{checkInputValidity}
\alias{checkInputValidity}
\title{Detectors Assignment Function}
\usage{
checkInputValidity(
  id = NULL,
  s,
  y,
  z,
  lowerHabCoords,
  upperHabCoords,
  localHabWindowIndices = NULL,
  numLocalHabWindows = NULL,
  trapCoords,
  localTrapsIndices = NULL,
  localTrapsNum = NULL,
  resizeFactor = 1,
  habitatGrid,
  habitatGridLocal = NULL,
  silent = FALSE,
  printReport = FALSE,
  pathReport = NULL
)
}
\arguments{
\item{data}{A \code{sf} object containing the detection data}

\item{detectors}{A \code{list} of sf objects containing the detector locations}

\item{subDetectors}{Optional; a \code{list} of sf objects containing the subdetectors locations (for PAB models only)}

\item{radius}{a \code{numeric} value denoting the maximum radius to consider for detector assignment. 
Any detection further than \code{radius}m from any detector will not be assigned to any detector. Instead it will get a 'NA'.}
}
\value{
A \code{sf} object with the x and y locations of the
 different detections with associated detector's Id
}
\description{
\code{checkInputValidity} identifies and assigns the closest detector to each detection in \code{data}.
}
