% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanRovBaseData.R
\name{cleanRovbaseData}
\alias{cleanRovbaseData}
\title{Data set clean-up.}
\usage{
cleanRovbaseData(
  species,
  years = NULL,
  data_dir = "./Data",
  output_dir = "./Data",
  Rmd_template = NULL,
  overwrite = FALSE
)
}
\arguments{
\item{species}{A \code{character} string with the name of the focal species
("bear", "wolf", or "wolverine").}

\item{years}{A \code{numeric} vector containing the years of interest. 
Only data for those years will be cleaned and returned.}

\item{data_dir}{the \code{path} pointing to the directory containing the raw 
data from Rovbase.}

\item{output_dir}{the \code{path} pointing to the directory where the cleaned 
data will be stored. The clean data, as well as a \code{.html} report describing 
the content of the clean data, will be placed in a folder with the species name,
inside a subfolder with the date of extraction of the raw Rovbase data.}

\item{Rmd_template}{the \code{path} to the \code{.rmd} template to be used for
cleaning the data. By default, the \code{.rmd} template provided with the 
\code{rovquantR} package is used.}

\item{overwrite}{A \code{logical} Whether previously existing clean data should
be overwritten or not (default = FALSE).}
}
\value{
This function returns:
\enumerate{
\item A \code{.RData} file with the clean NGS and dead recovery data objects
 for the species and period specified.
\item A \code{.html} report summarizing the data cleaning process. 
\item Additional \code{.png} images that can be reused somewhere else.
}
}
\description{
\code{cleanRovbaseData} calls a custom Rmarkdown template that identifies and loads
 the most recent RovBase data available on dropbox for the specified species
 and conducts a set of data cleaning steps that include:
 - removing un-identified samples
 - checking sex-assignment
 - removing flagged samples from RovData/Mike
}
\examples{
cleanRovbaseData("bear", 2012:2021)
cleanRovbaseData("Jerv", 2020)
cleanRovbaseData("Gray wolf", 2018:2022)

}
\author{
Pierre Dupont
}
