% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCleanEstimates.R
\name{getCleanEstimates}
\alias{getCleanEstimates}
\title{Print Bayesian model estimates in a clean way.}
\usage{
getCleanEstimates(x, moment = "mean", quantiles = c(0.025, 0.975))
}
\arguments{
\item{x}{a vector of posterior MCMC samples.}

\item{moment}{an \code{character string} denoting the type of metric to print; can be one of "mean" or "median".}

\item{quantiles}{a numeric vector of length 2 denoting the boundaries of the credible interval (e.g. c(0.025,0.975), the default value for the 95% credible interval).}
}
\value{
A formated character string with the mean (or median) and the associated credible interval in parentheses.
}
\description{
R utility function to print summary metrics from a vector of posterior samples.
}
\details{
The \code{getCleanEstimates} function is used to print the posterior mean (or median) of a vector of posterior MCMC samples, together with the credible interval of the specified width (in parentheses).
}
\author{
Pierre Dupont
}
