% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{GetDensity}
\alias{GetDensity}
\alias{quantileCpp}
\alias{fastIntMode}
\alias{extractUniquePositiveValues}
\alias{createTransitionMatrix}
\alias{GetSpaceUse}
\alias{GetDetectability_normal}
\alias{GetDetectability_mean}
\alias{GetTransitions}
\title{Density extraction}
\usage{
quantileCpp(x, q)

fastIntMode(x, narm = FALSE)

extractUniquePositiveValues(matrix)

createTransitionMatrix(values)

GetDensity(
  sx,
  sy,
  z,
  IDmx,
  aliveStates,
  regionID,
  probs = as.numeric(c(0.025, 0.975)),
  display_progress = TRUE,
  returnPosteriorCells = TRUE
)

GetSpaceUse(
  sx,
  sy,
  z,
  sigma,
  habitatxy,
  aliveStates,
  regionID,
  probs = as.numeric(c(0.025, 0.975)),
  display_progress = TRUE,
  returnPosteriorCells = TRUE
)

GetDetectability_normal(
  p0,
  sigma,
  habitatxy,
  detectorxy,
  regionID,
  probs = as.numeric(c(0.025, 0.975)),
  localDist = 100,
  display_progress = TRUE,
  returnPosteriorCells = FALSE
)

GetDetectability_mean(
  p0,
  sigma,
  habitatxy,
  detectorxy,
  display_progress = TRUE
)

GetTransitions(
  sx1,
  sy1,
  z1,
  sx2,
  sy2,
  z2,
  stateFrom,
  stateTo,
  rgmx,
  probs = as.numeric(c(0.025, 0.975)),
  display_progress = TRUE
)
}
\arguments{
\item{sx}{A numeric vector whose sample quantiles are wanted.}

\item{sy}{A logical, whether to remove NA values before calculation or not.}

\item{z}{A logical, whether to remove NA values before calculation or not.}

\item{aliveStates}{A logical, whether to remove NA values before calculation or not.}

\item{regionID}{A logical, whether to remove NA values before calculation or not.}

\item{probs}{A logical, whether to remove NA values before calculation or not.}

\item{display_progress}{A logical, whether to remove NA values before calculation or not.}

\item{returnPosteriorCells}{A logical, whether to remove NA values before calculation or not.}

\item{sigma}{A logical, whether to remove NA values before calculation or not.}

\item{habitatxy}{A logical, whether to remove NA values before calculation or not.}
}
\value{
Return_Description
}
\description{
The \code{getDensity}
}
\details{
The \code{getSpaceUse} function extracts density estimates for a set of grid cells based onindividual posterior space use distribution.
}
