% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{GetDensity}
\alias{GetDensity}
\alias{quantileCpp}
\alias{fastIntMode}
\alias{extractUniquePositiveValues}
\alias{createTransitionMatrix}
\alias{GetSpaceUse}
\alias{GetDetectability_normal}
\alias{GetDetectability_mean}
\alias{GetTransitions}
\title{MCMC output post-processing}
\usage{
quantileCpp(x, q)

fastIntMode(x, narm = FALSE)

extractUniquePositiveValues(matrix)

createTransitionMatrix(values)

GetDensity(
  sx,
  sy,
  z,
  IDmx,
  aliveStates,
  regionID,
  probs = as.numeric(c(0.025, 0.975)),
  display_progress = TRUE,
  returnPosteriorCells = TRUE
)

GetSpaceUse(
  sx,
  sy,
  z,
  sigma,
  habitatxy,
  aliveStates,
  regionID,
  probs = as.numeric(c(0.025, 0.975)),
  display_progress = TRUE,
  returnPosteriorCells = TRUE
)

GetDetectability_normal(
  p0,
  sigma,
  habitatxy,
  detectorxy,
  regionID,
  probs = as.numeric(c(0.025, 0.975)),
  localDist = 100,
  display_progress = TRUE,
  returnPosteriorCells = FALSE
)

GetDetectability_mean(
  p0,
  sigma,
  habitatxy,
  detectorxy,
  display_progress = TRUE
)

GetTransitions(
  sx1,
  sy1,
  z1,
  sx2,
  sy2,
  z2,
  stateFrom,
  stateTo,
  rgmx,
  probs = as.numeric(c(0.025, 0.975)),
  display_progress = TRUE
)
}
\arguments{
\item{sx}{A numeric vector whose sample quantiles are wanted.}

\item{sy}{A logical, whether to remove NA values before calculation or not.}

\item{z}{A logical, whether to remove NA values before calculation or not.}

\item{aliveStates}{A logical, whether to remove NA values before calculation or not.}

\item{regionID}{A logical, whether to remove NA values before calculation or not.}

\item{probs}{A logical, whether to remove NA values before calculation or not.}

\item{display_progress}{A logical, whether to remove NA values before calculation or not.}

\item{returnPosteriorCells}{A logical, whether to remove NA values before calculation or not.}

\item{sigma}{A logical, whether to remove NA values before calculation or not.}

\item{habitatxy}{A logical, whether to remove NA values before calculation or not.}
}
\value{
Return_Description
}
\description{
This set of \code{C++} functions are meant to facilitate the post-processing of SCR and OPSCR posterior samples in order to provide relevant information and derived quantities.
}
\details{
The \code{GetDensity} function extracts density estimates for a set of grid cells based on individual posterior activity center locations.
The \code{GetSpaceUse} function extracts density estimates for a set of grid cells based on individual posterior space use distribution.
The \code{GetDetectability_normal} function calculates detectability, i.e. the probability for an individual to be detected at least once for a set of grid cells based on posterior baseline detection probabilities and scale parameters. This function assumes a half-normal detection function.
The \code{GetDetectability_mean} function is a fast alternative of the \code{GetDetectability_normal} based on the posterior means of the parameters instead of the full posterior MCMC.
}
