% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{GetDensity}
\alias{GetDensity}
\title{AC-based density extraction}
\usage{
GetDensity(
  sx,
  sy,
  z,
  IDmx,
  aliveStates,
  regionID,
  probs = as.numeric(c(0.025, 0.975)),
  display_progress = TRUE,
  returnPosteriorCells = TRUE
)
}
\arguments{
\item{x}{A numeric vector whose sample quantiles are wanted.}

\item{narm}{A logical, whether to remove NA values before calculation or not.}
}
\description{
This function extract density estimates for a set of grid cells based on
individual posterior activty center coordinates.he sample mode.
}
