% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSInits.R
\name{getSInits}
\alias{getSInits}
\title{Generation of activity centers for SCR and OPSCR initial values.}
\usage{
getSInits(
  AllDetections = AllDetections,
  Id.vector = Id.vector,
  idAugmented = idAugmented,
  lowerCoords = lowerCoords,
  upperCoords = upperCoords,
  habitatGrid = habitatGrid,
  intensity = NULL,
  sd = 4,
  movementMethod = "dbernppACmovement_normal"
)
}
\arguments{
\item{AllDetections}{a dataframe.}

\item{Id.vector}{A character vector.}

\item{idAugmented}{A character string.}

\item{lowerCoords, upperCoords}{Matrices of lower and upper x- and y-coordinates of all habitat windows scaled to the habitat (see (\code{\link[nimbleSCR]{scaleCoordsToHabitatGrid}}).
One row for each window. Each window should be of size 1x1.}

\item{habitatGrid}{Matrix of habitat window indices. Cell values should correspond to the order of habitat windows in \code{lowerCoords} and \code{upperCoords}. 
When the habitat grid only consists of a single row or column of windows, an additional row or column of dummy indices has to be added because the \code{nimble} model code requires a matrix.}

\item{intensity}{A \code{numeric} vector.}

\item{sd}{Standard deviation of the isotropic bivariate normal distribution.}

\item{movementMethod}{a character denoting the nimbleSCR function to be used.}
}
\value{
This function returns a 2- or 3-dimensional array of initial AC locations.
}
\description{
R utility function to generate initial AC locations using compiled Nimble functions for efficiency.
}
\details{
The \code{getSInits} function is used in advance of model building.
}
\author{
Pierre Dupont
}
