% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeAugmentation.R
\name{MakeAugmentation}
\alias{MakeAugmentation}
\title{MakeAugmentation function}
\usage{
MakeAugmentation(y, aug.factor = NULL, aug.years = NULL, replace.value = NA)
}
\arguments{
\item{y}{A \code{Vector}, \code{Matrix} or \code{Array} object containing the individual detection histories.}

\item{aug.factor}{A \code{numeric} object defining the augmentation factor to be used.}

\item{replace.value}{A \code{numeric} object defining the value to be repeated for augmented individuals.}
}
\value{
A \code{Vector}, \code{Matrix} or \code{Array} object containing the augmented y.
}
\description{
\code{MakeAugmentation} increases the dimensions of an object along
 the individual and/or the year dimension. It returns a \code{Vector}, \code{Matrix} 
 or \code{Array} object with the expanded object.
}
