% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeDirectories.R
\name{makeDirectories}
\alias{makeDirectories}
\title{Prepare folder structure}
\usage{
makeDirectories(path = NULL, subFolders = NULL, show.dir = TRUE)
}
\arguments{
\item{path}{A path indicating where to create the different folders.}

\item{subFolders}{A \code{character vector} with the names of the subfolders to be created for the nimble input and output files.}

\item{show.dir}{A \code{logical}. (Optional; requires package  \emph{fs}) If TRUE, the directory structure created is displayed in a tree-like format.}
}
\value{
boolean invisible(FALSE) if nothing was created, invisible(TRUE) if folders were created in \emph{path}.
}
\description{
\code{makeDirectories} creates a set of folders to store all the necessary
data for the OPSCR ROvQuant analysis.
}
\examples{
\dontrun{makeDirectories()}

}
\author{
Pierre Dupont
}
