% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeRovquantData.R
\name{makeRovquantData}
\alias{makeRovquantData}
\title{RovQuant OPSCR data preparation.}
\usage{
makeRovquantData(
  species = c("bear", "wolf", "wolverine"),
  data_dir = "./Data",
  working_dir = NULL,
  years = NULL,
  sex = c("Hann", "Hunn"),
  aug.factor = 2,
  sampling.months = list(4, 5, 6, 7, 8, 9, 10, 11),
  habitat.res = 20000,
  buffer.size = 50000,
  max.move.dist = 250000,
  detector.res = 5000,
  subdetector.res = 1000,
  max.det.dist = 70000,
  resize.factor = 1,
  plot.check = FALSE,
  print.report = TRUE
)
}
\arguments{
\item{species}{A \code{character} denoting the species; can be any of "bear", "wolf" or "wolverine.}

\item{data_dir}{A \code{path} to the directory containing the clean Rovbase data,
as prepared by \code{cleanRovBaseData}.}

\item{working_dir}{A \code{path} to the directory for this analysis containing
the \code{nimbleInputFiles} folder to store the prepared data.}

\item{years}{A \code{list}  Whether dead recovery should be included (TRUE) or not(FALSE)}

\item{sex}{A \code{character} denoting the species; can be any of "bear", "wolf" or "wolverine.}

\item{aug.factor}{A \code{numeric} to the directory containing the clean Rovbase data,}

\item{sampling.months}{A \code{list}.}

\item{habitat.res}{A \code{numeric}  Whether dead recovery should be included (TRUE) or not(FALSE)}

\item{buffer.size}{A \code{numeric} denoting the species; can be any of "bear", "wolf" or "wolverine.}

\item{max.move.dist}{A \code{numeric}.}

\item{detector.res}{A \code{numeric} to the directory for this analysis containing}

\item{subdetector.res}{A \code{numeric}  Whether dead recovery should be included (TRUE) or not(FALSE)}

\item{max.det.dist}{A \code{numeric} denoting the species; can be any of "bear", "wolf" or "wolverine.}

\item{resize.factor}{A \code{numeric}.}

\item{plot.check}{A \code{logical}.}

\item{print.report}{A \code{logical}.}
}
\value{
This function returns:
\enumerate{
\item A \code{.RData} file with the clean NGS and dead recovery data objects
 for the species and period specified.
\item A \code{.html} report summarizing the data cleaning process. 
\item Additional \code{.png} images that can be reused somewhere else.
}
}
\description{
The \code{makeRovquantData} function calls a custom Rmarkdown template that identifies 
and loads the most recent Rovbase data available for the specified species 
and performs the OPSCR data preparation for model fitting. 
The data preparation is specific to each species and incorporates the
different developments from project RovQuant to allow fitting large-scale
SCR and OPSCR models (see Turek et al., 2021 <doi.org/10.1002/ecs2.3385>  for more details):
\enumerate{
\item Rescaling spatial coordinates to a grid for fast look-up assignment.
\item A local evaluation approach (see Milleret et al., 2019 <doi:10.1002/ece3.4751> for more details)
\item A sparse matrix representation of the observation data to reduce the size of objects to be processed.
\item An indicator to shortcut calculations for individuals unavailable for detection.
}
}
\author{
Pierre Dupont
}
