% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeRovquantData_bear.R
\name{makeRovquantData_bear}
\alias{makeRovquantData_bear}
\title{RovQuant OPSCR brown bear data preparation.}
\usage{
makeRovquantData_bear(
  data_dir = "./Data",
  working_dir = NULL,
  years = NULL,
  sex = c("Hann", "Hunn"),
  aug.factor = 2,
  sampling.months = list(4, 5, 6, 7, 8, 9, 10, 11),
  habitat.res = 20000,
  buffer.size = 50000,
  max.move.dist = 250000,
  detector.res = 5000,
  subdetector.res = 1000,
  max.det.dist = 70000,
  resize.factor = 1,
  plot.check = FALSE,
  print.report = TRUE
)
}
\arguments{
\item{data_dir}{A \code{path}.}

\item{working_dir}{A \code{path}.}

\item{years}{A \code{list}.}

\item{sex}{A \code{character}.}

\item{aug.factor}{A \code{Numeric}.}

\item{sampling.months}{A \code{list}.}

\item{habitat.res}{A \code{Numeric}.}

\item{buffer.size}{A \code{Numeric}.}

\item{max.move.dist}{A \code{Numeric}.}

\item{detector.res}{A \code{Numeric}.}

\item{subdetector.res}{A \code{Numeric}.}

\item{max.det.dist}{A \code{Numeric}.}

\item{resize.factor}{A \code{Numeric}.}

\item{plot.check}{A \code{Logical}.}

\item{print.report}{A \code{Logical}.}
}
\value{
A \code{html} report summarizing the data preparation process
Additional \code{.png} images that can be reused somewhere else.
}
\description{
\code{makeRovquantData_bear} calls a custom Rmarkdown template that identifies
 and loads the most recent Rovbase data available for the specified 
 species and conducts a set of data cleaning steps that include:
 - removing un-identified samples
 - checking sex-assignment
 - removing flagged samples from RovData/Mike
}
\author{
Pierre Dupont
}
