% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeRovquantData_bear.R
\name{makeRovquantData_bear}
\alias{makeRovquantData_bear}
\title{RovQuant OPSCR brown bear data preparation.}
\usage{
makeRovquantData_bear(
  data_dir = "./Data",
  working_dir = NULL,
  years = NULL,
  sex = c("Hann", "Hunn"),
  aug.factor = 2,
  sampling.months = list(4, 5, 6, 7, 8, 9, 10, 11),
  habitat.res = 20000,
  buffer.size = 50000,
  max.move.dist = 250000,
  detector.res = 5000,
  subdetector.res = 1000,
  max.det.dist = 70000,
  resize.factor = 1,
  plot.check = FALSE,
  print.report = TRUE
)
}
\arguments{
\item{dead_recoveries}{A \code{DataFrame} object containing the raw dead recoveries data.}

\item{country_polygon}{A \code{SpatialPointsDataFrame} with Country polygon for correct assignment of samples to countries}

\item{threshold_month}{A \code{Numeric} with initial month of the biological year: 1:January...12=December. all samples with months<threshold get year-1. so they get similar year.}

\item{keep_dead}{A \code{logical}  Whether dead recovery should be included (TRUE) or not(FALSE)}
}
\value{
A \code{.RData} file with the clean NGS and dead recovery data objects
for the species and period specified.
A \code{html} report summarizing the data cleaning process
Additional \code{.png} images that can be reused somewhere else.
}
\description{
\code{makeRovquantData_bear} calls a custom Rmarkdown template that identifies
 and loads the most recent Rovbase data available for the specified 
 species and conducts a set of data cleaning steps that include:
 - removing un-identified samples
 - checking sex-assignment
 - removing flagged samples from RovData/Mike
}
\author{
Pierre Dupont
}
