% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeRovquantData.R, R/makeRovquantData_bear.R
\name{makeRovquantData}
\alias{makeRovquantData}
\alias{makeRovquantData_bear}
\title{Data preparation.}
\usage{
makeRovquantData(
  species = c("bear", "wolf", "wolverine"),
  data_dir = "./Data",
  working_dir = NULL,
  years = NULL,
  sex = c("Hann", "Hunn"),
  aug.factor = 2,
  sampling.months = list(4, 5, 6, 7, 8, 9, 10, 11),
  habitat.res = 20000,
  buffer.size = 50000,
  max.move.dist = 250000,
  detector.res = 5000,
  subdetector.res = 1000,
  max.det.dist = 70000,
  resize.factor = 1,
  plot.check = FALSE,
  print.report = TRUE
)

makeRovquantData_bear(
  data_dir = "./Data",
  working_dir = NULL,
  years = NULL,
  sex = c("Hann", "Hunn"),
  aug.factor = 2,
  sampling.months = list(4, 5, 6, 7, 8, 9, 10, 11),
  habitat.res = 20000,
  buffer.size = 50000,
  max.move.dist = 250000,
  detector.res = 5000,
  subdetector.res = 1000,
  max.det.dist = 70000,
  resize.factor = 1,
  plot.check = FALSE,
  print.report = TRUE
)
}
\arguments{
\item{species}{A \code{character} denoting the species; can be any of "bear", "wolf" or "wolverine.}

\item{data_dir}{A \code{path} to the directory containing the clean Rovbase data,
as prepared by \code{\link{cleanRovBaseData}}.}

\item{working_dir}{A \code{path} to the directory for this analysis containing
the \code{nimbleInputFiles} folder to store the prepared data.}

\item{dead_recoveries}{A \code{DataFrame} object containing the raw dead recoveries data.}

\item{country_polygon}{A \code{SpatialPointsDataFrame} with Country polygon for correct assignment of samples to countries}

\item{threshold_month}{A \code{Numeric} with initial month of the biological year: 1:January...12=December. all samples with months<threshold get year-1. so they get similar year.}

\item{keep_dead}{A \code{logical}  Whether dead recovery should be included (TRUE) or not(FALSE)}
}
\value{
This function returns:
\enumerate{
\item A \code{.RData} file with the clean NGS and dead recovery data objects
 for the species and period specified.
\item A \code{.html} report summarizing the data cleaning process. 
\item Additional \code{.png} images that can be reused somewhere else.
}

A \code{.RData} file with the clean NGS and dead recovery data objects
for the species and period specified.
A \code{html} report summarizing the data cleaning process
Additional \code{.png} images that can be reused somewhere else.
}
\description{
\code{makeRovquantData} calls a custom Rmarkdown template that identifies 
and loads the most recent Rovbase data available for the specified species 
and performs the OPSCR data preparation for model fitting. 
The data preparation is specific to each species and incorporates the
different developments from project RovQuant to allow fitting large-scale
SCR and OPSCR models (see Turek et al., 2021 <doi.org/10.1002/ecs2.3385>  for more details):
\enumerate{
\item Rescaling spatial coordinates to a grid for fast look-up assignment.
\item A local evaluation approach (see Milleret et al., 2019 <doi:10.1002/ece3.4751> for more details)
\item A sparse matrix representation of the observation data to reduce the size of objects to be processed.
\item An indicator to shortcut calculations for individuals unavailable for detection.
}

\code{makeRovquantData_bear} calls a custom Rmarkdown template that identifies
 and loads the most recent Rovbase data available for the specified 
 species and conducts a set of data cleaning steps that include:
 - removing un-identified samples
 - checking sex-assignment
 - removing flagged samples from RovData/Mike
}
\examples{
makeRovquantData_bear(2012:2021)


makeRovquantData_bear(2012:2021)

}
\author{
Pierre Dupont
}
