% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeSearchGrid.R
\name{MakeSearchGrid}
\alias{MakeSearchGrid}
\title{Function to create detectors}
\usage{
MakeSearchGrid(
  data = data,
  resolution = resolution,
  div = 1,
  center = T,
  plot = TRUE
)
}
\arguments{
\item{data}{A \code{sp} object. can be a \code{SpatialPolygons} or \code{SpatialPoints} or \code{SpatialLines} or \code{raster}. Spatial Lines objects takes time to compute}

\item{resolution}{Numeric variable denoting the size of grid cells in units of (\code{polygon}).}

\item{div}{Numeric variable denoting the number of equally-sized subplots grid cell is to be divided into.}

\item{center}{Logical variable denoting whether nodes of the resulting search grid are to be centered within their respective subplots.}

\item{plot}{Logical for whether (\code{TRUE}) or not (\code{FALSE}) check plots are to be generated.}
}
\description{
\code{MakeSearchGrid} creates  detectors within a defined area (data) and 
resolution between detectors. Division precises if subspatial division should
 be performed (Following PAB model).
}
\author{
Richard Bischof, \email{richard.bischof@nmbu.no} and Pierre Dupont
}
