% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDensityMaps.R
\name{plotDensityMaps}
\alias{plotDensityMaps}
\title{Plot population density maps}
\usage{
plotDensityMaps(
  input,
  estimates,
  unit = 100,
  mask,
  background = NULL,
  type = c("time.series", "last.year", "summary"),
  path = getwd(),
  species = NULL,
  labels = NULL,
  x.labels = NULL,
  y.labels = NULL,
  caption = NULL,
  export.raster = TRUE,
  name = "UD_Density"
)
}
\arguments{
\item{input}{Raster used for extraction, as prepared by the \code{getDensityInput} function.}

\item{estimates}{Density estimates as provided by the \code{getDensity} or \code{getSpaceUse} functions.}

\item{unit}{Unit (in km-2) to be used for plotting density.}

\item{mask}{A \code{raster} object denoting the region for which density should be plotted.
All grid cells with values different from NA will be plotted.
Note that the extent and resolution of this raster must match the extent and resolution of the raster in \code{input}.}

\item{background}{A \code{sf} object of the overall region for which density should be plotted.}

\item{type}{A \code{character} string denoting the type of plot to be printed. Can be one of \code{"last.year"}, \code{"time.series"} or \code{"all"}.}

\item{path}{A \code{path} denoting where to save the density maps.}

\item{species}{(optional) A \code{character} string with the name of the species to be displayed (only used with \code{type = "summary"}).}

\item{export.raster}{A \code{logical}; whether to save density rasters as .tif files or not.}

\item{name}{A \code{character} string with the name to be used when saving .png and .tif files.}

\item{q95}{(optional) A \code{numeric} vector with the boundaries of the 95% credible interval for the population size estimates to be displayed (only used with \code{type = "summary"}).}
}
\value{
This function prints out \code{.png} image of density maps in the specified folder.
}
\description{
Function to plot (and print as \code{.pdf} files) population density maps from SCR or OPSCR models.
}
\details{
The \code{plotDensityMaps} function uses the outputs from the \code{getDensityInput} and \code{getDensity} functions as inputs.
}
\author{
Pierre Dupont
}
