% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processCodaOutput.R
\name{processCodaOutput}
\alias{processCodaOutput}
\title{Process coda output}
\usage{
processCodaOutput(x, DIC = F, params.omit = NULL, verbose = TRUE)
}
\arguments{
\item{x}{A \code{mcmc.list} object.}

\item{DIC}{A \code{logical} if DIC statistics should be computed or not.}

\item{params.omit}{A \code{vector} of parameters for which statistics should not be computed.}

\item{verbose}{A \code{logical}. If TRUE, messages are displayed.}
}
\value{
A \code{list} with sims lists, parameter summary values and rhat values.
}
\description{
\code{processCodaOutput} Process a coda output to make it readable. Those functions
 are from https://github.com/kenkellner/jagsUI/blob/master/R/processoutput.R
}
