% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rovquantR_data.R
\docType{data}
\name{REGIONS}
\alias{REGIONS}
\title{COUNTIES}
\format{
## `REGIONS`
A simple feature collection with 38 features and 4 fielsd#'
}
\source{
<https://www.nmbu.no/en/research/projects/rovquant>
}
\usage{
REGIONS
}
\description{
These data sets describe Scandinavian counties.

`COUNTIES`, `REGIONS`, `COUNTRIES`, `COUNTRIESWaterHumans` and `GLOBALMAP` are simple feature (sf) objects.
`habitatRasterResolution` and `habitatRasters` are raster objects.
}
\examples{
data(REGIONS)
}
\keyword{datasets}
