% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runMCMCbites.R
\name{runMCMCbites}
\alias{runMCMCbites}
\title{Run bite-size NIMBLE MCMC algorithms}
\usage{
runMCMCbites(
  mcmc,
  model = NULL,
  conf = NULL,
  bite.size,
  bite.number,
  path,
  save.rds = FALSE
)
}
\arguments{
\item{mcmc}{a \code{NIMBLE MCMC algorithm}. See details.}

\item{bite.size}{an \code{integer} denoting the number of MCMC iterations in 
each MCMC bite.}

\item{bite.number}{an \code{integer} denoting the number of MCMC bites to be run.}

\item{path}{a \code{character string} of the path where MCMC bite outputs are
to be saved as .RData files (when using \code{runMCMCbites}) or looked for 
(when using \code{collectMCMCbites}).}

\item{burnin}{an \code{integer} denoting the number of MCMC bites to be removed 
as burn-in.}

\item{pattern}{a \code{character string} denoting the name of the object containing
MCMC samples in each .R Data bite file.}

\item{param.omit}{a \code{character vector} denoting the names of parameters 
that are to be ignored when combining MCMC bites.}

\item{progress.bar}{a \code{logical value}. if TRUE, a separate progress bar is 
printed for each MCMC chain.}
}
\description{
\code{runMCMCbites} is a wrapper R function to run a compiled NIMBLE model in 
multiple small size bites. This reduces memory usage and allows saving MCMC 
samples on the fly.

\code{restartMCMCbites} is a wrapper R function to restart and extend a previous
MCMC run, starting from the last point it was saved (as .rds file). In addition to
the saved model and MCMC states, this requires the MCMC configuration, nimble model
and MCMC to be built as in the original MCMC run.
By default, \code{restartMCMCbites} identifies and uses the last .rds file as a starting point.
the \code{path.rds} option allows to use a specific .rds file as starting point instead.

\code{collectMCMCbites} is a convenience R function to combine multiple bites
from the same MCMC chain into a single \code{mcmc} object (as defined in the 
\code{coda} package). If multiple chains were run, MCMC bites are combined 
chain-wise and compiled into a \code{mcmc.list} object.
}
